refactor: diary like test 분리
